final online model with db

+++ mysql/create_table.sql
@@ -6,7 +6,7 @@ GRANT ALL PRIVILEGES ON `default`.* TO '%'@'%';
 GRANT ALL PRIVILEGES ON `default`.* TO 'confluent2'@'%';
 
 CREATE TABLE ticker_ethusdt (
-    closeTime BIGINT,
+    closeTime INTEGER,
     count DOUBLE,
     firstId DOUBLE,
     highPrice DOUBLE,
@@ -14,7 +14,7 @@ CREATE TABLE ticker_ethusdt (
     lastPrice DOUBLE,
     lowPrice DOUBLE,
     openPrice DOUBLE,
-    openTime DOUBLE,
+    openTime INTEGER,
     priceChange DOUBLE,
     priceChangePercent DOUBLE,
     quoteVolume DOUBLE,
